finishing 1/2 of appove decline mechanism

+++ ghost/core/core/admin/app/components/gh-reports-list-item.js
@@ -3,7 +3,7 @@ import {action} from '@ember/object';
 import {formatPostTime} from 'ghost-admin/helpers/gh-format-post-time';
 import {inject as service} from '@ember/service';
 import {tracked} from '@glimmer/tracking';
-import {task} from 'ember-concurrency';
+import {task,timeout} from 'ember-concurrency';
 export default class GhReportsListItemComponent extends Component {
   @service feature;
   @service session;
@@ -11,7 +11,8 @@ export default class GhReportsListItemComponent extends Component {
   @service store
 
   @tracked isHovered = false;
-  @tracked isDeleted = false;
+  @tracked isExisted = true;
+  @tracked isSaved = true;
   @action
   mouseOver() {
       this.isHovered = true;
@@ -22,20 +23,45 @@ export default class GhReportsListItemComponent extends Component {
       this.isHovered = false;
   }
   @task(function* () {
-      alert(1);
+      let {report} = this.args;
+      let url = this.store.createRecord('reporturl',{
+        id:report.id,
+        report_link : report.report_link,
+        type: report.type,
+        created_date: report.created_date,
+      })
+      if(!url.isDeleted){
+        alert("succesfully add report url")
+        url.save()
+        yield report.destroyRecord()
+        if(!report.isDeleted)
+        {
+          alert("delete report error ")
+        }
+        else{
+          yield timeout(1000)
+          alert("succesfully add report")
+          this.isExisted = false;
+        }
+      }
+      else{
+        alert("error")
+      }
+
   })
       approveTask;
 
   @task(function* () {
       let {report} = this.args;
-      report.destroyRecord()
+      yield report.destroyRecord()
       if(!report.isDeleted)
       {
         alert("delete error")
       }
       else{
+        yield timeout(1000)
         alert("succesfully")
-        this.isDeleted = true;
+        this.isExisted = false;
       }
   })
       declineTask;
